+++ PipesRUs/src/pipesrus/PriceEngine/PipeTypeThree.java
@@ -11,7 +11,7 @@ package pipesrus.PriceEngine;
  */
 public class PipeTypeThree extends PipeTypeTwo
 {
-    public PipeTypeThree(float length, float diameter, PipeGrade grade, boolean chemResist)
+    public PipeTypeThree(double length, double diameter, PipeGrade grade, boolean chemResist)
     {
         super(length, diameter,grade, chemResist);
         this._colour = PipeColour.TWO_COLOURS;
